Revert "Refactored code to be more modular"
This reverts commit 3090dafd303b721c3ff9f7b65bd9f05d2d1132fc.
