--- src/main/java/Program.java
@@ -1,185 +1,0 @@
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.util.ArrayList;
-import java.util.Map;
-import java.util.function.Consumer;
-import java.util.function.Function;
-
-import duke.tasks.Deadline;
-import duke.tasks.Event;
-import duke.tasks.Task;
-import duke.tasks.ToDo;
-
-public final class Program {
-  public final static void main(String[] vargs) throws IOException {
-      String logo = " ____        _        \n"
-              + "|  _ \\ _   _| | _____ \n"
-              + "| | | | | | | |/ / _ \\\n"
-              + "| |_| | |_| |   <  __/\n"
-              + "|____/ \\__,_|_|\\_\\___|\n";
-      System.out.println("Hello from\n" + logo);
-
-      final BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
-      final ArrayList<Task> tasks = new ArrayList<>();
-
-      final Consumer<Task> addTask = (task) -> {
-          tasks.add(task);
-          System.out.format("Added %s to the list!\nYou now have %d tasks\n", task.toString(), tasks.size());
-      };
-
-      final Function<String, Integer> intParser = (value) -> {
-          int index = Integer.parseInt(value);
-          if (index < 1 || index > tasks.size()) {
-              throw new NumberFormatException();
-          }
-          return index;
-      };
-
-      final Map<String, Consumer<String[]>> funcMap = Map.of(
-      "bye", (args) -> {
-          System.out.println("Ok bye bye!");
-          System.exit(0);
-      },
-      "list", (args) -> {
-          if (tasks.size() == 0) {
-              System.out.println("No stored tasks!");
-          } else {
-              System.out.format("A total of %d tasks\n", tasks.size());
-              for (int i = 0; i < tasks.size(); i++) {
-                  System.out.format("%d. %s\n", i + 1, tasks.get(i).toString());
-              }
-          }
-      },
-      "mark", (args) -> {
-          try {
-              if (args.length == 1) {
-                  System.out.println("Needed an index for mark");
-                  return;
-              }
-
-              int index = intParser.apply(args[1]);
-              tasks.get(index - 1).setDone(true);
-              System.out.format("Marked this as done!\n%s\n", tasks.get(index - 1).toString());
-          } catch (NumberFormatException e) {
-              System.out.println("Invalid index!\n");
-          }
-      },
-      "unmark", (args) -> {
-          if (args.length == 1) {
-              System.out.println("Needed an index for unmark");
-              return;
-          }
-          try {
-              int index = intParser.apply(args[1]);
-              tasks.get(index - 1).setDone(false);
-              System.out.format("Marked this as not done!\n%s\n", tasks.get(index - 1).toString());
-          } catch (NumberFormatException e) {
-              System.out.println("Invalid index!");
-          }
-      },
-      "todo", (args) -> {
-          if (args.length == 1) {
-              System.out.println("Expected a task!");
-              return;
-          }
-
-          String taskStr = joiner(args, 1, args.length);
-          Task task = new ToDo(taskStr);
-          addTask.accept(task);
-      },
-      "deadline", (args) -> {
-          int index = -1;
-          for(int i = 1; i < args.length; i++) {
-              if (args[i].equalsIgnoreCase("/by")) {
-                  index = i;
-                  break;
-              }
-          }
-
-          if (index == -1) {
-              System.out.println("Expected a /by directive!");
-              return;
-          } else if (index == 1) {
-              System.out.println("Expected a task!");
-              return;
-          } else if (index == args.length - 1) {
-              System.out.println("Expected a time after /by!");
-              return;
-          }
-
-          String taskStr = joiner(args, 1, index);
-          String time = joiner(args, index + 1, args.length);
-          Task task = new Deadline(taskStr, time);
-          addTask.accept(task);
-      },
-      "event", (args) -> {
-          int fromIndex = -1, toIndex = -1;
-          for(int i = 1; i < args.length; i++) {
-              if (args[i].equalsIgnoreCase("/from")) {
-                  fromIndex = i;
-              } else if (args[i].equalsIgnoreCase("/to")) {
-                  toIndex = i;
-              }
-
-              if (fromIndex != -1 && toIndex != -1) break;
-          }
-
-          if (fromIndex == -1) {
-              // No /from provided
-              System.out.println("Expected a /from directive!");
-              return;
-          } else if (toIndex == -1) {
-              // No /to provided
-              System.out.println("Expected a /to directive!");
-              return;
-          } else if (fromIndex == 1) {
-              // From was the first keyword after event
-              System.out.println("Expected a task!");
-              return;
-          } else if (fromIndex > toIndex) {
-              System.out.println("Expected /from to come before /to!");
-              return;
-          } else if (toIndex == args.length - 1) {
-              System.out.println("Expected a time after /to!");
-              return;
-          } else if (toIndex - fromIndex == 1) {
-              System.out.println("Expected a time after /from!");
-              return;
-          }
-
-          String taskStr = joiner(args, 1, fromIndex);
-          String fromStr = joiner(args, fromIndex + 1, toIndex);
-          String toStr = joiner(args, toIndex + 1, args.length);
-      
-          Task task = new Event(taskStr, fromStr, toStr);
-          addTask.accept(task);
-      },
-      "delete", (args) -> {
-        try {
-              if (args.length == 1) {
-                  System.out.println("Needed an index for delete");
-                  return;
-              }
-
-              int index = intParser.apply(args[1]);
-              Task removed = tasks.remove(index - 1);
-              System.out.format("Removed this task!\n%s\n", removed.toString());
-          } catch (NumberFormatException e) {
-              System.out.println("Invalid index!");
-          }
-      });
-
-      while (true) {
-          String input = reader.readLine();
-          String[] tokens = input.split(" ");
-
-          if (funcMap.containsKey(tokens[0])) {
-              funcMap.get(tokens[0]).accept(tokens);
-          } else {
-              System.out.format("Unknown command '%s'\n", input);
-          }
-          
-      }
-  }
-}
