+++ src/main/java/Task.java
@@ -1,4 +1,3 @@
-package duke.tasks;
 /**
  * Abstract class from which all other tasks are derived
  */
