+++ src/main/java/ToDo.java
@@ -1,4 +1,3 @@
-package duke.tasks;
 /**
  * A task without any associated times
  */
